on off command working

+++ app/Services/DispanserService.php
@@ -198,12 +198,12 @@ class DispanserService extends ServiceProvider
         }				
 
         $stopCommand = Process::where('type_id', 5)->where('pfc_id', $pfc_id)->count();
-        if($stopCommand != 0){      	  	    
-            Process::where('type_id', 1)->where('pfc_id', $pfc_id)->delete();  	 
-			//if(Process::where('type_id', 1)->where('pfc_id', $pfc_id)->count() != 0){
+        if($stopCommand != 0){      	  	     	 
+			if(Process::where('type_id', 1)->where('pfc_id', $pfc_id)->count() != 0){
+				Process::where('type_id', 1)->where('pfc_id', $pfc_id)->delete(); 
 				socket_close($socket);
 				exit();
-			//}
+			}
 			return false;
         }
         $stopCommand = Process::where('type_id', 6)->where('pfc_id', $pfc_id)->count();
